requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision
ultralytics
streamlit
opencv-python-headless
matplotlib
reportlab
pillow
numpy
pandas
